DISPATCH-56: Unifiy C and python management agents.
There are 2 management agents.
Previously the C agent was at the default address $management, python agent at $management2
This commit completes all missing features of the python agent and makes it the default on $management.
The old C agent is still available on $cmanagement, but it is deprecated and will be removed.

- Added C entity types to the python schema: router.link, router.address, router.node, connection, allocator
- Update attribute values from C in python agent.
- Implemented get-operations, get-attributes, get-mgmt-nodes
- Added optional file:line in logging.

git-svn-id: https://svn.apache.org/repos/asf/qpid/dispatch/trunk@1632696 13f79535-47bb-0310-9956-ffa450edef68