DISPATCH-56: Consistent naming style for managment attribute and entity names.
Converted all entity and attribute names for AMQP management to lower camelCase
style like this: fooBar Previously there was a mix of styles for attribute
names: foo_bar, foo-bar and fooBar.  The fooBar style is used by the AMQP
management draft spec. and the Qpid Java and C++ brokers so is the most
consistent choice.

Configuration files are another way to specify management entity
attributes. Dispatch still accepts the foo-bar style in configuration files for
backwards compatibility, but it also accepts the fooBar style.

git-svn-id: https://svn.apache.org/repos/asf/qpid/dispatch/trunk@1633704 13f79535-47bb-0310-9956-ffa450edef68