DISPATCH-72: Sporadic core dumps in dispatch tests.
Added missing lock scopes around calls to python code, this resoloved the
sporadic core dumps (over 1000 test runs overnight with no failures)

Initially I speculated that the problem was that we don't lock the GIL as recommended by
the python C API docs https://docs.python.org/2/c-api/init.html#non-python-created-threads.
Instead we use a separate mutex in the C code.

I attempted to use the GIL but that caused deadlocks. They appeared to be
associated with locking a python mutex in the management agent - this causes
python to unlock and relock the GIL. That allows other threads into the python
code which somehow causes deadlock - I am not sure exactly why.

I believe it is OK to use the separate C mutex *provided* we never create
threads in python code.  If we do that then I think we will need to re-visit the
GIL and figure out how to lock it correctly.

git-svn-id: https://svn.apache.org/repos/asf/qpid/dispatch/trunk@1633814 13f79535-47bb-0310-9956-ffa450edef68