DISPATCH-80: Management schema: implement annotations and entity inheritance as per AMQP management WD 09
Renamed "includes" as "annotations" - The AMQP management 09 draft introduces annotations
which serve exactly the same function as our includes, so renamed according to the standard.

Introduced entity inheritance and created the following hierarchy among entity types:

"entity": the base of all entity types.
- "configurationEntity": base for all types that hold configuration information.
- "operationalEntity": base for all types that hold operational information.

"Configuration" information is supplied in advance and expresse intent. For
example "I want the router to listen on port N", or, "I want the router to
connect to this address" All the entities that can be used in the configuration
file are "configurationEntity".

"Operational" information reflects the actual state of the router at a given
moment in time.  For example, "how many addresses are presently active on the
router?" All the entities queried by qdstat are "operationalEntity".

The two types are often related, for example "I want an on-demand connection to
this address" and "is there presently a live outgoing connection to this
address?"  As a matter of good practice we separate the two kinds of
entities. For example "listener" and "connector" are configuration entities
expressing the intent to make or receive connectoins. The "connection" entity
holds live operational information about an actual connection.

- Fixed man pages and configuration to use new hierarchy.
- Fixed use of "connection" as both annotation and entity type name.
- Implemented GET-ANNOTATIONS and GET-TYPES to return annotations and bases.

git-svn-id: https://svn.apache.org/repos/asf/qpid/dispatch/trunk@1643319 13f79535-47bb-0310-9956-ffa450edef68