DISPATCH-56: Improve qdmanage tool, add man page generation for qdmanage and qdstat.
Got rid of JSON maps/lists on qdmanage command line. Use simple name=value arguments for attributes.
Still allow JSON with the --stdin flag.

Man page generation, from the doc/man/README:

Command line tools --help option provides a very brief usage and list of options.

Man pages should be more detailed, but we do not want to repeat the information
from --help as it will easily become out of date.

To solve this we generate man pages using the  help2man utility.
For each tool there should be a tool.htm with extra details for the man page.
help2man will combine this with the --help otuput of the tool to create the full man page.

See the help2man macro in CMakeLists.txt and the help2man documentation for more details.

For building distributions when help2man is not available, the generated man page for
each tool is checked in as "tool.8.in". When help2man is available, running make
will warn if the checked in version is not  up to date with the generated version.

To update the man page you MUST use help2man, do not edit the checked in file
as your edits will be lost.

git-svn-id: https://svn.apache.org/repos/asf/qpid/dispatch/trunk@1644318 13f79535-47bb-0310-9956-ffa450edef68