DISPATCH-85: Remove all direct printing to stdout and stderr.
The only thing qdrouterd was printing to stderr was the debugging information
about final status of allocators. This can't be logged normally because the
logging system is already finalized at that point.

Solution: a new container configuration attribute debugDump. If set to a file
name, allocator info is written to that file during shutdown.

git-svn-id: https://svn.apache.org/repos/asf/qpid/dispatch/trunk@1644825 13f79535-47bb-0310-9956-ffa450edef68