NO-JIRA: Avoid use of GNU extension ## comma deletion.
The following idiom is GNU C specific:
-#define log(fmt, ...)
-            log_impl(fmt , ##__VA_ARGS__);

The use of ## after a comma deletes the comma if __VA_ARGS__ is empty, so the above works
even if there is no argument after fmt.

The following is more portable:

+#define log(...)
+            log_impl(__VA_ARGS__);

Now fmt is included in the variadic argument list so it is never empty.  This is
slightly less self-descriptive since it doesn't correspond as well to the
signature of log_impl(fmt, ...), but it is standard C99.

git-svn-id: https://svn.apache.org/repos/asf/qpid/dispatch/trunk@1645070 13f79535-47bb-0310-9956-ffa450edef68