DISPATCH-87: Double load of libqpid-dispatch.so
We use ctypes to call from python into the qpid-dispatch.so library.
In some situations (running a build exe where there was an installed dispatch),
the exe could open the build library but ctypes could open the installed library
causing mayhem and core dumps.

Fixed the code to use dlopen(RTLD_NOLOAD) in C to get a handle to the library loaded by C,
and then use the handle rather than the library name to open the library in ctypes
which ensures the same library.

NOTE: When running qdrouterd from a build you still need to use the right
environment (config.sh) to pick up the correct python libraries (correct
PYTHONPATH and internal libraries) or things will still go wrong.

git-svn-id: https://svn.apache.org/repos/asf/qpid/dispatch/trunk@1645976 13f79535-47bb-0310-9956-ffa450edef68