DISPATCH-26: Router documentation - use pandoc to generate router book in PDF and HTML.
If you install pandoc ("yum install pandoc" on yum-based systems) then make will
create PDF and HTML versions of the dispatch router book.

The source is still in markdown format, but note pandoc uses a different table
extension syntax so check an example in the existing .md files or the pandoc
documentation.

Did some edits to the existing book sources
- fleshed out the intro
- updated comments on management
- fixed up tables

git-svn-id: https://svn.apache.org/repos/asf/qpid/dispatch/trunk@1647638 13f79535-47bb-0310-9956-ffa450edef68