DISPATCH-121: Workaround to force cmake re-configure.
reconfigure.in is a workaround to force cmake re-configuration. For example,
we use GLOB to collect .h files for install and apidoc, so if you _remove_ a
.h file it won't trigger automatic re-configure and everybody's builds will
fail till they run cmake manually.

If you do check in such a change, increase the number in this file by 1.
That will force automatic re-configure and everybody will be happy.

git-svn-id: https://svn.apache.org/repos/asf/qpid/dispatch/trunk@1664743 13f79535-47bb-0310-9956-ffa450edef68