NO-JIRA: run_system_tests.py doesn't correctly locate config files.
run_system_tests.py was using the script location __file__ to find SSL certs and other config.
That works if the script is called with an absolute path (e.g. from ctest) but not if it
is called with a relative path (e.g. manually)

Changed the tests to use system_test.__file__ to find config. An imported module always has
an absolute path regardless of how the script is called.

git-svn-id: https://svn.apache.org/repos/asf/qpid/dispatch/trunk@1670985 13f79535-47bb-0310-9956-ffa450edef68