DISPATCH-219: System test security configuration for installed/built tests.
Got rid of cmake substitutions in python test files: On test errors, a python
IDE will zap to the relevant line in the *substituted copy* not the real python
source. Fixes accidentally made to the copy will be lost when source is copied
again.

Got rid of cmake substituted SASL configuration files. The system tests are
installed and can be run on an installation as well as in the build, so they
can't have build-specific paths in them. (Unit tests that are not installed
don't have this restriction)

Instead generate the SASL config files at run time in the test's local working
directory. The system_test.py framework ensures this is always an isolated
scratch directory used only by that test module.