Fix memory leak in router_config.c
The buffers that the composed field body is made out of are not owned by either the field iterator or the parse
that is run over them, so they need to be independently freed, but the publishing method can't free them because
they are referenced by the action which has been published to be run later.  We need to free them once the action
is run.  But, the only reference we had to them was inside of a field iter inside of a parse.  These can't assume
that they will free the underlying memmory because the specifically don't own it.  So, passed along a separate
reference to the underlying buffers so that the action handler could explicitly free the memory after it is used.
This is a little ugly in that the one client that is not using buffers and doesn't need to pass along anything to
free now has to pass along an empty buffer list.  Seemed like the best solution to me, but other possibilities do
exist.