DISPATCH-901: authorization from remote auth service
If the client specifies its desire for the ADDRESS-AUTHZ capacbility, the authorization service, if it supports this, will return a
set of permissions in the properties of the open frame. The properties will have an address-authz key, whose value is a map of address (or wildcard pattern) to an array of permissions.
The only permissions recognised at present by this patch are 'send' and 'recv'.