DISPATCH-993: Get test to complete in when messages are released
This commit is the heart of the Jira fix.

If enough messages are released then the test never sends enough
messages to trigger the spurious link deletion.
The send count is now (n_messages + n_released).

Test 01 debug reports the number and time elapsed for leading
released messages in the message stream.