DISPATCH-1318: Fix MobileAddress test and test framework issues
When the test detects an error print all the details and not just a
one line error that gets clobbered by the test framework.

Don't count receiver settlements as if they are sender settlements.

Print details of client activity with microsecond timestamps that may
be correlated with router logs.

Allow MODIFIED settlement state in addition to RELEASED for messages that
get bounced because of the closed receiver.