DISPATCH-1359: DISPATCH-480: Variable lengths for self test timeouts
DISPATCH-1359 is trying to shorten the default test timeout from 1500
to 300 seconds. It further makes a distinction between _fast_ self tests
that would benefit from a shorter timeout and _slow_ self tests that
should get a longer timeout.

DISPATCH-480 is requesting a greater timeout for slow self tests in
certain environments.

This patch addresses both jiras. It sets the default short timeout to
10 seconds and the default long timeout to 300 seconds. Then it lets
a user specify the short and long self test timeout values on the
cmake command line:

  cmake -DTEST_TIMEOUT_SHORT=20 -DTEST_TIMEOUT_LONG=600 ...

This closes #521