DISPATCH-1423: restore original mcast credit replenish behavior
Prior to DISPATCH-1266 credit was issued for multicast senders
unconditionally.  DISPATCH-1266 changed this to only grant credit when
subscribers were present.  This broke DISPATCH-779.

This patch fixes this by issuing credit on multicast links regardless
of the presence of subscribers.  Note well that unsettled mcast
messages that are successfully forwarded do not have their credit
replenished until after the receivers settle the message.