DISPATCH-1384 - Disable inlining of functions that are mocked in unit_tests
The timer_test.c relies on being able to provide its own dummy versions
of `qd_timer_now` and `qd_server_timeout`. That is not possible if these
functions are inlined in the .dynlib. Clang on macOS inlines them,
GCC on Linux doesn't (that's why the tests pass on Linux)