DISPATCH-1427: Add optional per-vhost-group connection limits
With these settings a vhost may be configured so that users in
a vhost group may receive different limits than the vhost has
overall. The original motivation was to allow administrators
more connections per user and per host than everyone else but
the limits may be applied to raise or lower the bounds for any
group.