DISPATCH-1536: Microsecond timestamp utility for self-tests
This commit pulls the Logger class out of system_tests_edge_router and
places it into system_tests for easy access.

An example consumer of Logger is the MobileAddressTest. That test creates
a logger and loads it with facts as the test progresses but it doesn't print
anything unless there is a test failure.

By default Logger does not print to the console and it saves each log
line as it goes. When Logger is initialized either of these defaults
may be overridden.

How to print to the console with a timestamp:

    from system_test import Logger
    ...
    self.log2 = Logger(print_to_console=True, save_for_dump=False)
    self.log2.log("Help I'm a Rock")