NO-JIRA: Instrument MessageRouteAbortTest; use new test Logger
This patch adds debugability to system_tests_delivery_abort.
 - Every test message is a different length
 - Test message content is unique - a string of the message length
 - Logging statements track test progress

A major frustration is a failure like DISPATCH-1559: some test fails
intermittently and the results of the assertion that fails the test is
*truncated*. It is an even harder problem when the failure is on some
remote CI system over which one has no control.

This patch is essentially the debug scaffolding that was required to
debug DISPATCH-1559. This patch uses the Logger added in DISPATCH-1536.

The log holds a narrative of exactly when (to the uS) each message was
sent or aborted from the client's view. Armed with this one can then
look into the router logs to see AMQP progress and diagnose the fault.