DISPATCH-1678: fix misleading ASAN tracebacks for pooled items
This patch updates the alloc_pool.c debug code to free any "leaked"
items to prevent ASAN from printing tracebacks for them.  These
tracebacks are misleading since ASAN only tracks the initial heap
malloc, not the last time the item was allocated from the pool.

Add a list of alloc_pool.c types that are known/expected to leak. If
the alloc_pool.c debug code finds a leak for a type which is NOT on
this list the router will abort on shutdown (debug build only).

Update the .travis.yml to clean up warnings and include a test run of
the default build case.