DISPATCH-1679: Add sys_thread_self()
sys_thread_self() provides an interface for accessing the sys_thread_t
pointer for the currently running thread.  This replaces the old
sys_thread_id() function which required the thread to already have its
sys_thread_t pointer, which basically defeats the purpose.  In
addition the return value of sys_thread_id() was pthread specific and
should only be compared using pthread_equal().  sys_thread_self()
return value is a plain pointer that can be compared in line.

This closes #760