DISPATCH-1679: Make qd_timer_cancel and qd_timer_free synchronous
This patch makes the qd_timer_cancel() and qd_timer_free() routines
block if the timer callback is currently running on another thread.
This ensures that the callback is not executing when these calls
return.

It includes a fix to the HTTP1 adapter to avoid freeing its timer
while holding the adaptor lock.

This closes #1170