DISPATCH-1676: PolicyWarnings test process shutdown issue
This test start and stops a router so quickly that it exposes process
handling issues in the test framework. That problem has nothing to do
with Policy.

The test was written originally to prove that python Policy code could
write log warnings withtout crashing the router. That issue was fixed
long ago.