DISPATCH-1690: clean up last inbound message on link close
If an incoming link fails/closes before the current inbound message
has been forwarded to the core the current inbound message is not
freed.  This patch adds a weak reference (safe pointer) in the link to
the current inbound message.  This reference is cleared once the
message has been delivered to the core.  Should the link close prior
to delivery this weak reference is used to clean up the message.

This closes #766