DISPATCH-1585: Define vhost aliases for configuation sharing
Given a vhost definition:

  vhost {
     hostname: example.com
     aliases: example.org, example.net
     ... }

Vhost _aliases_ are alternative literal hostnames or patterns that direct
the router to use the settings in this vhost. Alias hostnames that match an
incoming connection will use the settings defined in the contained vhost.
In a multi-tenant configuration a connection to a vhost alias will use the
base vhost hostname for the tenant namespace. In this example if a connection
is directed to vhost `example.org` then the settings from the base vhost
hostname `example.com` will apply and `example.com` will be the tenant
namespace.

Vhost `hostname` and `aliases` settings from all vhosts are shared in a
single list and must be unique.

This closes #770