DISPATCH-1505: Qdstat shows policy and vhost entities
  -p --policy   - base policy settings and status
  --vhosts      - vhost configuration with no group data
  --vhostgroups - vhost group configuration
  --vhoststats  - per-vhost statistics, per-vhost/user connections info

Policy is present even if it is not specified in the router config file.
Vhosts are present only if enabled in base policy.

The vhostgroups display is kind of ugly with superwide rows and
does not lend itself to the tabular format.  Similar to --log, it
is excluded from --all-entities and printed only on demand.

This closes #774