DISPATCH-1721: Qdstat table vhostgroups: too wide; missing default settings
At a minimum the table is about 200 columns wide. If large lists of user
names and/or sources and targets are specified the table is so wide
that it becomes unreadable. This patch breaks up vhostgroups displays:

 * for all vhost and groups: a single table of boolean and integer settings
 * for each (vhost, vhostusergroup) pair: an individual table of strings

Entity management does not return appropriate defaults for table
vhostUserGroupSettings. To remedy that:

 * Fetch the schema from the target router and use those defaults.

 * Enforce a 'false' default on schema booleans that are not required
   and have no default.

This closes #780