DISPATCH-1737: Log config file json as-modified when json parse fails
When config file json parsing fails then the parser prints a line and
column number of the offending input. Since the original input is modified
by the config process the reported line and column numbers are not useful.

This patch logs the input that json was given when parsing fails.
The parser failure line and column numbers directly relate to logged
json input lines and syntax errors may be identifed with some authority.

This closes #808