DISPATCH-1751: Rework how AMQP session incoming-window is derived
Proton allows specification of a session 'capacity'. Initially the
incoming-window will be (capacity / max-frame-size), defining at most
how many max-size transfers will definitely be accepted on the session.

Dispatch listener config defines a maxFrameSize and a maxSessionFrames
the product of which is equal to the capacity to be configured in Proton.

Dispatch vhostUserGroup policy defines a maxFrameSize and a
maxSessionWindow. The maxSessionWindow is passed directly to Proton
as the capacity.

The interface to Proton defines capacity with type 'size_t'. It is possible
to have a (maxFrameSize * maxSessionFrames) product be larger than
that allowed by a size_t on a 32-bit system. Perform the capacity
calculations using 64-bit integers before applying the results and
sending them to Proton.