DISPATCH-1846: Debugging the A-B-C hang
This patch is not meant to be merged. DEBUG ONLY!

Refresh: A TCP listener on INTA can't connect to a TCP server
on INTC. INTA-INTA works, and INTA-INTB works.

The TCP self test is trimmed to just the three interior routers.
Then bunches of log statements are sprinkled around to track what
the routers are doing.

Note that log statements in router-proper code are from log
source 'TCP_TEST' at critical level. They appear in Scraper logs
as coming from container 'TCP_TEST'.  This is all just for debug.