DISPATCH-1895: Modify tcp self test not to use http listeners
As noted in jira comment, the test failure is the http-libwebsockets
code doing a wild 64-bit realloc. The failure has nothing to do with
the tcp test itself and should be directed to a new jira blaming
the libwebsockets code.

That said, the tcp test is using http listeners purely for easier
test debugging. The http listeners may be removed with no loss of
test functionality and that is what this patch does.