DISPATCH-1952: fix qd_bitmask_t message streaming leak
This patch moves the point where the message annotations are updated
to after the early return tiggered by the check for Q2/streaming.
This avoids the qd_bitmask_t leak by allocating it after the
Q2/streaming check passes.  It also avoids potentially doing the
annotation computation multiple times for the same message.

This closes #595