DISPATCH-1981: TCP self test to force Q2 flow control
The echo server is rewritten to generate Q2 holdoff more reliably.

The self test uses the connection-stall feature of the the echo
server. The server does not read the message content, beyond normal
TCP window and python prefetch, so that the router gets into a
Q2 block state.

This closes #1060