DISPATCH-1878: Handle half-closed TCP connections DISPATCH-2043: Various TCP Adaptor logging improvements
* monitor and control sequence of half closed events
* raw read caches data if read_close before stream message created
* raw write defers EOS write_close until stream message created

This version leaks buffers at the rate of about 15 per connection pair.
The suspect is that proton never emits a PN_RAW_CONNECTION_READ after
emitting CLOSED_READ and this code never gets to reclaim the buffers.
That said, this code could have a bug in that area as well.