DISPATCH-2091: Ignore RX window when raw connection is write_closed
If the adaptor receives an EOS then it calls pn_raw_write_close. This
sends a FIN to the raw connection peer.

If the adaptor then continues to honor the RX window closure then it
never reads data from the raw connection. Further, when the incoming
TCP window is full then the TCP connection never returns the client's
FIN and the connection is stuck open.

This patch causes the adaptor to keep reading from an connection after
a write_close has been effected.