DISPATCH-2091: Ignore RX window when raw connection is write_closed
When TCP adaptor receives an EOS in the incoming streaming link then
it calls pn_raw_write_close to the corresponding PN raw connection.
This sends a FIN to the raw connection peer.

If the adaptor then continues to honor the RX window closure then it
never reads data from the raw connection that now has a close in
progress. Further, when the incoming TCP window is full then the
raw connection never returns the client's FIN and the connection
is stuck open.

This patch causes the adaptor to override backpressure flow control
and to keep reading from a raw connection after a write_close has been
effected.

This closes #1177