NO-JIRA: remove KEYS file.
The proper KEYS file to use for verifying downloaded artifacts
can be found here: https://downloads.apache.org/qpid/KEYS

This is explained on the Qpid project website.

Do NOT include a KEYS file in the dispatch tarball since an attacker
can spoof the one in the tarfile itself.

This closes #1215