DISPATCH-1274: avoid sharing mutex between timer and server
Historically the qd_server_t and the timer subsystem have shared a
single mutex. There is no technical reason for this as each use the
lock to protect entirely separate critial sections.

Recent analysis of lock usage shows that the timer subsystem makes
heavy use of the lock.  This patch introduces separate locks as to
avoid any contention between the qd_server_t and timer operations.