DISPATCH-1487: ensure message context is set up for locally generated messages
This patch modifies how locally generated messages are
initialized. The associated message context is now initialized
with the proper locations of any message sections present. This
will allow the router annotations to be composed on message send,
rather than requiring the message creator to manually set the
proper trace, ingress, etc. fields when the message is created.

This patch also removes the plethora of "qd_message_compose_X()"
functions and replaces them with a single general purpose message
constructor.