DISPATCH-2357: fix(Dockerfile): switch to using Rocky Linux 8 as CentOS 8 is EOL
We need to use supported CentOS. In newer CentOS Streams, the PowerTools repo was renamed to crb (same name as in RHEL).
Using Rocky Linux 8 seems most sensible, as we are not ready to upgrade all the way to CentOS 9 or 10.

The Dockerfile was initially intended as a help for development, not really a release artifact.
If a development Dockerfile is needed, it can be added in the future.

Meanwhile, using development versions of software at random places, which then becomes part of our release tar.gz, creates bad impression and should be eliminated.

As a compromise, the Dockerfile can use the contents of the local directory for Dispatch sources.
In a released tar.gz, that would be the released content, so we're fine there.